backend/requirements.txt
fastapi
uvicorn[standard]

selenium
webdriver-manager

undetected-chromedriver

python-multipart

python-dotenv

setuptools

jinja2>=3.1.4
markupsafe>=2.1.5
starlette>=0.40.0
